refactor(odoo): organize action implementations
