+++ integrations/odoo/src/odoo-client/actions/implementations/contacts/index.ts
@@ -1,0 +1,6 @@
+export { createContact } from './createContact'
+export { deleteContacts } from './deleteContacts'
+export { getContactFields } from './getContactFields'
+export { getContacts } from './getContacts'
+export { searchContacts } from './searchContacts'
+export { updateContacts } from './updateContacts'
