+++ integrations/odoo/src/odoo-client/actions/implementations/current-user/getCurrentUser.ts
@@ -1,4 +1,4 @@
-import { wrapAction } from '../action-wrapper'
+import { wrapAction } from '../../action-wrapper'
 
 export const getCurrentUser = wrapAction(
   { actionName: 'getCurrentUser', errorMessage: 'Failed to get current Odoo user' },
