+++ integrations/odoo/src/odoo-client/actions/implementations/current-user/index.ts
@@ -1,0 +1,1 @@
+export { getCurrentUser } from './getCurrentUser'
