+++ integrations/odoo/src/odoo-client/actions/implementations/index.ts
@@ -1,16 +1,6 @@
-import { createContact } from './createContact'
-import { createLead } from './createLead'
-import { deleteContacts } from './deleteContacts'
-import { deleteLeads } from './deleteLeads'
-import { getContactFields } from './getContactFields'
-import { getContacts } from './getContacts'
-import { getCurrentUser } from './getCurrentUser'
-import { getLeadFields } from './getLeadFields'
-import { getLeads } from './getLeads'
-import { searchContacts } from './searchContacts'
-import { searchLeads } from './searchLeads'
-import { updateContacts } from './updateContacts'
-import { updateLeads } from './updateLeads'
+import { createContact, deleteContacts, getContactFields, getContacts, searchContacts, updateContacts } from './contacts'
+import { getCurrentUser } from './current-user'
+import { createLead, deleteLeads, getLeadFields, getLeads, searchLeads, updateLeads } from './leads'
 import * as bp from '.botpress'
 
 export default {
