+++ integrations/odoo/src/odoo-client/actions/implementations/leads/getLeadFields.ts
@@ -1,4 +1,4 @@
-import { wrapAction } from '../action-wrapper'
+import { wrapAction } from '../../action-wrapper'
 
 export const getLeadFields = wrapAction(
   { actionName: 'getLeadFields', errorMessage: 'Failed to get Odoo lead fields' },
