+++ integrations/odoo/src/odoo-client/actions/implementations/leads/getLeads.ts
@@ -1,4 +1,4 @@
-import { wrapAction } from '../action-wrapper'
+import { wrapAction } from '../../action-wrapper'
 
 export const getLeads = wrapAction(
   { actionName: 'getLeads', errorMessage: 'Failed to get Odoo leads' },
