+++ integrations/odoo/src/odoo-client/actions/implementations/leads/index.ts
@@ -1,0 +1,6 @@
+export { createLead } from './createLead'
+export { deleteLeads } from './deleteLeads'
+export { getLeadFields } from './getLeadFields'
+export { getLeads } from './getLeads'
+export { searchLeads } from './searchLeads'
+export { updateLeads } from './updateLeads'
