+++ integrations/odoo/src/odoo-client/actions/implementations/leads/searchLeads.ts
@@ -1,4 +1,4 @@
-import { wrapAction } from '../action-wrapper'
+import { wrapAction } from '../../action-wrapper'
 
 export const searchLeads = wrapAction(
   { actionName: 'searchLeads', errorMessage: 'Failed to search Odoo leads' },
